+++ apps/docs/package.json
@@ -35,14 +35,18 @@
     "@radix-ui/react-tooltip": "catalog:",
     "@radix-ui/react-use-controllable-state": "catalog:",
     "@shikijs/transformers": "catalog:",
-    "@tailwindcss/typography": "catalog:",
     "babel-plugin-react-compiler": "catalog:",
     "class-variance-authority": "catalog:",
+    "hastscript": "catalog:",
     "clsx": "catalog:",
     "cmdk": "catalog:",
     "lucide-react": "catalog:",
     "next": "catalog:",
+    "next-mdx-remote": "catalog:",
     "next-themes": "catalog:",
+    "rehype-autolink-headings": "catalog:",
+    "rehype-slug": "catalog:",
+    "remark-gfm": "catalog:",
     "postcss": "catalog:",
     "postcss-import": "catalog:",
     "react": "catalog:",
