+++ apps/docs/tsconfig.json
@@ -4,6 +4,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
+    "**/*.mdx",
     ".next/types/**/*.ts",
     "types/**/*.d.ts"
   ],
