+++ apps/docs/types/mdx.d.ts
@@ -1,0 +1,5 @@
+declare module '*.mdx' {
+  import type { MDXProps } from 'mdx/types';
+
+  export default function MDXContent(props: MDXProps): JSX.Element;
+}
