+++ package.json
@@ -73,7 +73,6 @@
       "@svgr/plugin-jsx": "8.1.0",
       "@svgr/plugin-svgo": "8.1.0",
       "@tailwindcss/postcss": "4.1.13",
-      "@tailwindcss/typography": "0.5.19",
       "@trivago/prettier-plugin-sort-imports": "6.0.0",
       "@types/bun": "1.3.2",
       "@types/hast": "3.0.4",
@@ -93,10 +92,15 @@
       "eslint-plugin-react-hooks": "7.0.0",
       "fast-deep-equal": "3.1.3",
       "hast-util-to-html": "9.0.5",
+      "hastscript": "9.0.0",
       "lru_map": "0.4.1",
       "lucide-react": "0.545.0",
       "next": "16.0.7",
+      "next-mdx-remote": "5.0.0",
       "next-themes": "0.4.6",
+      "rehype-autolink-headings": "7.1.0",
+      "rehype-slug": "6.0.0",
+      "remark-gfm": "4.0.0",
       "postcss": "8.5.6",
       "postcss-html": "1.7.0",
       "postcss-import": "16.1.1",
